+++ src/cp/actions/fields/PerformRowAction.php
@@ -8,11 +8,9 @@
 
 namespace flipbox\force\cp\actions\fields;
 
-use Craft;
 use craft\base\ElementInterface;
 use flipbox\ember\actions\traits\Manage;
 use flipbox\force\criteria\SObjectCriteria;
-use flipbox\force\db\SObjectFieldQuery;
 use flipbox\force\fields\actions\SObjectRowActionInterface;
 use flipbox\force\fields\SObjects;
 use flipbox\force\Force;