+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Changelog
 =========
 ## Unreleased
 ### Fixed
-- CP notifications were not being displayed after field actions occured
+- CP notifications were not being displayed after field actions occurred
+- Instance where an environment 
 
 ## 1.0.0-beta.1 - 2018-04-26
 ### Fixed